+++ src/Elements/Section.php
@@ -2,36 +2,16 @@
 
 namespace Tlr\Tables\Elements;
 
-use Tlr\Tables\Elements\Interfaces\Element;
 use Tlr\Tables\Elements\Interfaces\HasChildren;
+use Tlr\Tables\Elements\Interfaces\RowInterface;
+use Tlr\Tables\Elements\Interfaces\SectionInterface;
 use Tlr\Tables\Elements\Row;
 use Tlr\Tables\Elements\Traits\Attributable;
 use Tlr\Tables\Elements\Traits\Classable;
 
-class Section implements Element, HasChildren
+abstract class Section implements SectionInterface, HasChildren
 {
     use Attributable, Classable;
-
-    /**
-     * The parent table
-     *
-     * @var \Tlr\Tables\Elements\Table
-     */
-    protected $table;
-
-    /**
-     * The table section element
-     *
-     * @var string
-     */
-    protected $element;
-
-    /**
-     * The cell element for this section
-     *
-     * @var string
-     */
-    protected $cellElement;
 
     /**
      * The section's rows
@@ -40,30 +20,13 @@ class Section implements Element, HasChildren
      */
     protected $rows = [];
 
-    public function __construct(Table $table, string $element = 'tbody', string $cellElement = 'td')
-    {
-        $this->table = $table;
-        $this->element = $element;
-        $this->cellElement = $cellElement;
-    }
-
-    /**
-     * Make a new row
-     *
-     * @return \Tlr\Tables\Elements\Row
-     */
-    public function row() : Row
-    {
-        return $this->rows[] = $row = new Row($this);
-    }
-
     /**
      * Get the next row from the one given. Creates a new row if it's the last
      * row
      *
-     * @return \Tlr\Tables\Elements\Row
+     * @return \Tlr\Tables\Elements\Interfaces\RowInterface
      */
-    public function nextRow(Row $current) : Row
+    public function nextRow(Row $current) : RowInterface
     {
         $index = array_search($current, $this->rows, true);
 
@@ -72,36 +35,6 @@ class Section implements Element, HasChildren
         }
 
         return $this->rows[$index + 1] ?? $this->row();
-    }
-
-    /**
-     * Get the parent table
-     *
-     * @return \Tlr\Tables\Elements\Table
-     */
-    public function table() : Table
-    {
-        return $this->table;
-    }
-
-    /**
-     * Get the element
-     *
-     * @return string
-     */
-    public function getElement() : string
-    {
-        return $this->element;
-    }
-
-    /**
-     * Get the cell element
-     *
-     * @return string
-     */
-    public function getCellElement() : string
-    {
-        return $this->cellElement;
     }
 
     /**