+++ tests/integration/IntegrationTableTest.php
@@ -20,43 +20,33 @@ class IntegrationTableTest extends TestCase
         $this->assertEquals('<table><tbody><tr><td></td></tr></tbody></table>', $table->render());
     }
 
-    public function testBasicTableWithContentElement()
+    public function testBasicTableWithBodyAndAllCellElements()
     {
-        $table = new Table();
-        $table->body()->row()->cell('Foo');
+        $table = new Table;
+        $table->body()->row()->cell()->content('Foo');
+        $table->body()->row()->linkCell('foo.com')->content('bar');
+        $table->body()->row()->imageCell('foo.jpg');
 
-        $this->assertEquals('<table><tbody><tr><td>Foo</td></tr></tbody></table>', $table->render());
+        $this->assertEquals('<table><tbody><tr><td>Foo</td></tr><tr><td><a href="foo.com">bar</a></td></tr><tr><td><img src="foo.jpg" /></td></tr></tbody></table>', $table->render());
     }
 
-    public function testAllFeatures()
+    public function testBasicTableWithAllThreeSections()
     {
-        $table = (new Table())->class('ui table')->attribute('id', 'test-table');
+        $table = new Table;
+        $table->header()->row()->cell()->content('Foo');
+        $table->body()->row()->cell()->content('Bar');
+        $table->footer()->row()->cell()->content('Baz');
 
-        $row = $table->header()->row();
-        $row->cell()->spanColumns(2)->content('Name');
-        $row->cell()->spanRows(2)->content('Age');
+        $this->assertEquals('<table><thead><tr><th>Foo</th></tr></thead><tbody><tr><td>Bar</td></tr></tbody><tfoot><tr><th>Baz</th></tr></tfoot></table>', $table->render());
+    }
 
-        $row = $table->header()->row();
-        $row->cell()->content('First Name');
-        $row->cell()->content('Last Name');
+    public function testToStringRendersTheTable()
+    {
+        // This isn't really a unit test...
+        $table = new Table;
 
-        $expectedHeader = '<thead><tr><th colspan=2>Name</th><th rowspan=2>Age</th></tr><tr><th>First Name</th><th>Last Name</th></tr></thead>';
+        $table->class('foo bar');
 
-        $table->body()->row()
-            ->cell('Jeremy')->cell('Christmas')->cell(42)
-        ->row()
-            ->cell('Jeremy')->cell('Clarkson')->cell('>42');
-
-        $expectedBody = '<tbody><tr><td>Jeremy</td><td>Christmas</td><td>42</td><td>Jeremy</td><td>Clarkson</td><td>&gt;42</td></tr></tbody>';
-
-        $table->footer()->row()->cell('<a href="/new">New</a>')->raw();
-        $table->footer()->row()->cell()->wrapContent('<a href="/home">', '<Home>', '</a>');
-
-        $expectedFooter = '<tfoot><tr><th><a href="/new">New</a></th></tr><tr><th><a href="/home">&lt;Home&gt;</a></th></tr></tfoot>';
-
-        $this->assertEquals(
-            sprintf('<table class="ui table" id="test-table">%s%s%s</table>', $expectedHeader, $expectedBody, $expectedFooter),
-            $table->render()
-        );
+        $this->assertEquals($table->render(), (string)$table);
     }
 }